+++ examples/basic-aws-integration/outputs.tf
@@ -19,16 +19,36 @@ output "replicator_aws_iam_keys" {
   sensitive   = true
 }
 
+output "replicator_management_account_role_arn" {
+  description = "Amazon Resource Name (ARN) of Management Account Role for replicator"
+  value       = module.meshplatform.replicator_management_account_role_arn
+}
+
+output "replicator_automation_account_role_arn" {
+  description = "Amazon Resource Name (ARN) of Automation Account Role for replicator"
+  value       = module.meshplatform.replicator_automation_account_role_arn
+}
+
 output "replicator_privileged_external_id" {
   value       = module.meshplatform.replicator_privileged_external_id
   description = "Replicator privileged_external_id"
   sensitive   = true
 }
 
-output "kraken_aws_iam_keys" {
-  value       = module.meshplatform.kraken_aws_iam_keys
-  description = "You can access your credentials when you execute `terraform output kraken_aws_iam_keys` command"
+output "meshstack_access_role_name" {
+  value       = module.meshplatform.meshstack_access_role_name
+  description = "The name for the Account Access Role that will be rolled out to all managed accounts."
+}
+
+output "metering_aws_iam_keys" {
+  value       = module.meshplatform.metering_aws_iam_keys
+  description = "You can access your credentials when you execute `terraform output metering_aws_iam_keys` command"
   sensitive   = true
+}
+
+output "cost_explorer_management_account_role_arn" {
+  description = "Amazon Resource Name (ARN) of Management Account Role for replicator"
+  value       = module.meshplatform.cost_explorer_management_account_role_arn
 }
 
 output "cost_explorer_privileged_external_id" {